plugins/proj/server/pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "proj-mcp-server"
version = "0.23.8"
description = "MCP server for project lifecycle management"
requires-python = ">=3.12"
dependencies = [
    "mcp>=1.2.0",
    "pyyaml>=6.0",
    "gitpython>=3.1",
]

[project.scripts]
proj-server = "server.main:main"

[tool.hatch.build.targets.wheel]
packages = ["server"]

[dependency-groups]
dev = [
    "ruff>=0.14",
    "basedpyright>=1.2",
    "pre-commit>=3.7",
]
test = [
    "pytest>=8.0",
    "pytest-cov>=7.0",
    "pytest-xdist>=3.5",
    "pytest-asyncio>=0.23",
    "pytest-mock>=3.14",
    "hypothesis>=6.100",
    "freezegun>=1.5",
]
lint = [
    "bandit>=1.7",
    "pip-audit>=2.8",
]
docs = [
    "pdoc>=14.0",
]

[tool.ruff]
line-length = 100
target-version = "py312"

[tool.ruff.lint]
select = ["F", "E", "W", "I", "B", "C4", "UP", "SIM", "RUF", "TCH", "PTH", "S"]
ignore = ["S603", "S607"]

[tool.ruff.lint.per-file-ignores]
"tests/*" = ["S101", "S108", "TC001", "TC003"]

[tool.basedpyright]
typeCheckingMode = "strict"
pythonVersion = "3.12"

[tool.pytest.ini_options]
testpaths = ["tests"]
addopts = ["-v", "--tb=short", "--cov=server", "--cov-fail-under=72", "-n", "auto"]

[tool.coverage.run]
branch = true
omit = ["*/tests/*", "server/main.py", "server/cli.py"]
